Fix nav staying visible bug
Previously when clicking on the menu icon on a
small screen, the menu would drop down due to the
"visible" class being added. This class would
stay present on screen resize, causing the menu
to still be visible when going back to a smaller
screen. Now the class is removed on larger screens.